Initial Commit for GPS with Micro
This code is used with an Adafruit Ultimate GPS and an Arduino Micro. It
takes GPS data from the module, and feeds it into the Micro to be
parsed. It outputs the Time, Date, and the Latitude and Longitude to the
Serial Monitor, so it needs to be connected to a computer to view the
results